organize machineconfig tests
